🚚 (oxlint): rename `config` entrypoint to `utils`
